# services/lakehouse/openapi/lakehouse.yaml
openapi: 3.0.0
info:
  title: Lakehouse API
  description: |
    Asset storage and retrieval API for document management.
    Search PDFs by content, view files for vision analysis, process with PyMuPDF.

    This is an internal API that agents can discover and use like any external REST API.

    ## Lakehouse as a Compute Cache for Large API Responses

    The Lakehouse is not just document storage. It is a persistence layer for large
    structured blobs that are expensive to re-fetch from source APIs (Gmail, Slack,
    GitHub, Jira, etc.).

    ### The Core Pattern: Fetch → Blob → Ingest → Reuse

    1. **Fetch wide** — use the source API's native query language as a pre-filter.
       Pull raw results (emails, issues, messages, records) in batch using asyncio.gather()
       with a semaphore for concurrency.

    2. **Build a structured blob** — serialize all records into a single .txt or .json
       file using a consistent delimiter format:

       ```
       --- RECORD ---
       ID: {source_id}
       Source: {api_name}
       Date: {timestamp}
       Subject/Title: {label}
       Body:
       {content capped at 600 chars}
       ```

    3. **Ingest once** — POST /api/v1/assets/ingest (base64-encoded). The asset_id
       is your durable handle. Store it. The blob now survives container expiry.

    4. **Reuse across sessions** — any future execute_code() call retrieves the full
       blob via get_asset_text(asset_id) in a single API call, loading it directly
...

...

...

...

...

...

...
servers:
...
paths:
...
  /api/v1/assets/{asset_id}:
    get:
      operationId: get_asset_external
      summary: Get asset metadata (external)
      parameters:
        - name: asset_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Asset metadata
        '404':
          description: Not found
    put:
      operationId: update_asset_external
      summary: Replace asset content in-place
      description: |
        Replace the stored bytes of an existing asset. The S3 object is
        overwritten at the same key, metadata (content_type, size, checksum,
        extracted_text) is refreshed, and thumbnails are regenerated. All
        provenance metadata (created_at, user_id, session_id, tags, source)
        is preserved. Use this to edit an asset's content without changing
        its asset_id.
      parameters:
        - name: asset_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - base64_content
              properties:
                base64_content:
                  type: string
                  description: Base64-encoded replacement file bytes
                filename:
                  type: string
                  description: Optional new filename (keeps existing if omitted)
                tenant_id:
                  type: string
                  description: Optional tenant override (defaults to JWT tenant)
      responses:
        '200':
          description: Asset updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetResult"
        '400':
          description: Missing or invalid base64_content
        '401':
          description: Unauthorized
        '404':
          description: Asset not found
        '500':
          description: S3 upload failed
    delete:
      operationId: delete_asset_external
      summary: Delete asset (external)
      parameters:
        - name: asset_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
        '404':
          description: Not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        User JWT. Required on every authenticated route.
        Carries `sub` (user_id), `email`, and other identity claims.
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: |
        Service-to-service admin key (`MCP_ADMIN_KEY`). Optional.
        When present alongside `BearerAuth`, grants admin-level trust on
        top of the user JWT identity.
  schemas:
    AssetResult:
      type: object
      properties:
        asset_id:
          type: string
        filename:
          type: string
        content_type:
          type: string
        size_bytes:
          type: integer
        checksum_sha256:
          type: string
        s3_key:
          type: string
        thumbnail_url:
          type: string
          nullable: true
        snippet:
          type: string
          nullable: true
          description: First ~200 chars of extracted text (text/PDF assets only)
        tenant_id:
          type: string
          nullable: true
        status:
          type: string
          enum: [complete]
        updated_at:
          type: string
          format: date-time
        page_count:
          type: integer
          description: Present only for PDF assets
security:
  - BearerAuth: []
  - BearerAuth: []
    AdminKey: []